// pc/src/app/types/common.type.ts
export interface ResponseBody {
    data: any;
    msg: string;
    status: number;
}
// ...
export interface BankTypes {
    bankId: number;
    bankName: string;
}

// 银行卡类型
export interface BankInfo {
    cardUsername: any;
    bankId: any;
    cardNum: any;
    cardAddress: any;
    password: any;
}